+++ packages/template-retail-react-app/app/components/product-item/index.test.js
@@ -9,6 +9,7 @@ import ProductItem from '@salesforce/retail-react-app/app/components/product-ite
 import {mockedCustomerProductListsDetails} from '@salesforce/retail-react-app/app/mocks/mock-data'
 import {renderWithProviders} from '@salesforce/retail-react-app/app/utils/test-utils'
 import {screen} from '@testing-library/react'
+import PropTypes from 'prop-types'
 
 jest.mock('@salesforce/commerce-sdk-react', () => {
     const originalModule = jest.requireActual('@salesforce/commerce-sdk-react')
@@ -24,16 +25,55 @@ beforeEach(() => {
 })
 
 jest.setTimeout(60000)
-const MockedComponent = () => {
-    const product = mockedCustomerProductListsDetails.data[0]
-    return <ProductItem product={{...product, productName: product.name}} />
+
+const mockProduct = {
+    ...mockedCustomerProductListsDetails.data[0],
+    productName: mockedCustomerProductListsDetails.data[0].name,
+    bonusProductLineItem: false,
+    quantity: 1
 }
 
-test('renders product item name, attributes and price', async () => {
-    renderWithProviders(<MockedComponent />)
+const mockBonusProduct = {
+    ...mockProduct,
+    bonusProductLineItem: true
+}
 
-    // look for the element that has sole product name
-    expect(await screen.getByText(/apple ipod nano$/i)).toBeInTheDocument()
-    expect(await screen.getByText(/color: green/i)).toBeInTheDocument()
-    expect(await screen.getByText(/memory size: 16 GB$/i)).toBeInTheDocument()
+const MockedComponent = ({
+    product = mockProduct,
+    onItemQuantityChange = async () => {},
+    showLoading = false
+}) => {
+    return (
+        <ProductItem
+            product={product}
+            onItemQuantityChange={onItemQuantityChange}
+            showLoading={showLoading}
+            primaryAction={<button>Primary Action</button>}
+            secondaryActions={<button>Secondary Action</button>}
+        />
+    )
+}
+
+MockedComponent.propTypes = {
+    product: PropTypes.object,
+    onItemQuantityChange: PropTypes.func,
+    showLoading: PropTypes.bool
+}
+
+describe('ProductItem Component', () => {
+    test('renders product item name, attributes, price and quantity picker', async () => {
+        renderWithProviders(<MockedComponent />)
+
+        expect(await screen.getByText(/apple ipod nano$/i)).toBeInTheDocument()
+        expect(await screen.getByText(/color: green/i)).toBeInTheDocument()
+        expect(await screen.getByText(/memory size: 16 GB$/i)).toBeInTheDocument()
+        expect(screen.queryByRole('spinbutton')).toBeInTheDocument()
+    })
+
+    test('renders bonus product without quantity picker', () => {
+        renderWithProviders(<MockedComponent product={mockBonusProduct} />)
+
+        expect(screen.getByText(/Quantity:/i)).toBeInTheDocument()
+        expect(screen.queryByRole('spinbutton')).not.toBeInTheDocument()
+    })
 })